Resolve props for commit phase lifecycles
